Refactor builder into options pattern (#94)
This change replaces the builder pattern with the more idiomatic
option one. Tests reflect the previous behavior with the new pattern,
but I took the chance to refactor `pkg/config/` to avoid exposing an
additional interface and the `Validate()` function, encapsulating the
responsibility of creating a correct config within the package. All
users of `config.Config` now assume the config is valid.

Fixes #81