Add `deadcode` linter check (#98)
This adds a new `task deadcode` task that errors out if it finds
unused code. The check is implemented in a safe way using `-test`
option, so that code that is tested but not currently used in the
application does not cause a failure.

The `lint.yml` github action was modified to run this check as well.