Add `registryName` path element to Upstream API (#106)
This change adds support for two sets of Upstream API, one under
`/registry/v0.1` providing an aggregated view of all servers coming
from any configured registry, and one under
`/registry/{registryName}/v0.1` providing a list of all servers for a
given registry name.

In order to maintain compliance with Upstream API, the path element
was added _before_ the version number, which is not very elegant, but
we favor compatibility over elengance in this case.

This is part of #38