Decouple sync co-ordinator from status persistence
Prior to this PR, the sync coordinator delegated storage of state to a
status persistence type. However, the coordinator itself made a bunch of
assumptions about how sync state was stored which are no longer valid
now that we need to support a database - for example, it maintained a
cache of statuses which is unsuitable once we support multiple
registries sharing a database.

This PR addresses this by splitting the co-ordinator into two types: a
RegistryStateService type which provides operations for getting and
setting sync state on a given registry, and the coordinator which deals
with checking each registry's state on a schedule and triggering the
sync process as needed.

There is not a lot of new logic in this PR, but some code looks quite
different as a result of the new interface which separates the two
concerns.