Add db-backed implementation of `RegistryService` (#161)
This change adds a db-backed implementation of the `RegistryService`
interface.

There are multiple public interfaces added in this change,
specifically

* `internal/service` which exposes the `RegistryService` interface
  with an `Option` type used to configure filtering and pagination of
  SQL statements.
* `internal/service/db` which exposes means to create a
  `RegistryService` backed by a database.

Testing is implemented as integration tests running containers in the
same fashion as the `internal/db/sqlc` package. This might be revised
in the future by using mocks, but (a) a mock based implementation is
not going to be more succinct, and (b) does its job of covering most
of the new code.

Co-authored-by: Don Browne <dmjb@users.noreply.github.com>