Provide DB backed implementation of RegistryStateService (#191)
This mirrors the previous PR which provided a file backed implementation. This will be wired in a future PR.

Note that I changed the DB schema to match the existing sync status structure to simplify implementation. I also added a uniqueness constraint since we now assume that there is only one row in registry_sync per registry.

This PR also renames the LOCAL enum type to MANAGED, and ensures that configurations cannot change the registry type.