Change PK constraint on mcp_server_package table
Previously, the PK was inadvertantly set to be the same as the foreign
key. This led to a constraint where only package could be associated
with a registry at once. This didn't show up in testing so far since our
sample data only includes a single package per MCP server.

This changes the PK to the triple of (server_id, registry_type,
package_identifier). This should provide a meaningful definition of
package identity for our needs.