Remove redundant columns from temporary tables
This change refactors some SQL statements involving temporary tables
removing columns and their types from the statement itself. The reason
is that it's tricky to detect when these columns get out of sync with
the actual table, and the compiled version of those statements still
contains the full list, so it's basically the best of both worlds!

I also took the chance to improve the db setup routine dropping
environment variables watched by `libpq`. The net result is that the
test db is much more similar to an actual production deployment and is
guaranteed to not use any configuration local to the developer's
machine.