Simplify co-ordination logic to allow scaling and API configuration
Previously the server would set up a goroutine per registry and wait
until a registry sync was ready to begin. The logic now works as
follows:

1) The server spawns a single goroutine for monitoring sync statuses.
2) It calls a new GetNextSyncJob method which checks to see if a
   registry is ready to sync, and locks it if ready. This method checks
   all available registries, ordered by last sync time ascending to
   avoid starvation.
3) The goroutine either carries out the sync job and goes back to sleep,
   or goes back to sleep if there's nothing to do.

The polling period is separate to the per-registry sync interval. It is
currently set to 2 mins plus a random +/-30 second jitter interval to
minimize the chance of multiple instances hitting the DB at the same
time.

Advantages of the new approach:

1) Less code/complexity than previous implementation.
2) Allows horizontal scaling - multiple registries can share sync jobs
   without any additional co-ordination.
3) Allows us to implement endpoints to add/remove registries: the
   previous logic assumed that the set of registries was hardcoded at
   application startup.