Change PK of mcp_registry, take #2
The previous definition of the PK did not consider the transport type.
In practice, if an MCP server supports multiple transports, the upstream
registry will contain multiple packages which are identical except for
the value of the transport field (note that transport type is not an
array in the upstream schema, just a single string).

This also drops registry type from the PK definition since it is
unlikely that we will ever have two packages with the same identifier
but different registries.