Add `Kind` to entries exported from kubernetes
This change adds the Kubernetes `Kind` to metadata exported as part of
the MCP Server entries created in the Registry when auto-discovery is
active.

This change also modifies the structure of metadata making it a
structured object, following the same convention we follow in upstream
registry JSON format.

Finally, this also fixes a small bug that caused the Registry Server
to fail when creating a folder when DB storage is configured.

Fixes #279