Remove factory for RegistryDataProvider
The code previously had a factory for the RegistryDataProvider interface
which had stubbed out logic for returning an implementation of this type
backed by a database. In practice, this interface was not useful when we
added the DB, and so only the file-based implementation is needed.

This PR:

1) Gets rid of the factory. The application wireup code calls the
   factory method of the file-based implementation directly when wiring
   up a file-backed instance of the server.
2) Moves the file-based implementation and the interface under the
   inmemory package to make it clear that it is only relevant for the
   file-based storage logic.

In future, we may refactor the code to remove this interface entirely.