Simplify application wireup
Prior to this change, we had three different interfaces which DB or file based implementations. Each one had different dependencies which were specific to either the DB or to the file. This led to some complicated wiring in the app builder.

This code puts the creation of these types behind an interface, and an instance of that interface is selected based on the config (in other words, using the Abstract Factory pattern). This makes the builder logic easier to follow, and prevents the possibility of mixing DB and file implementations.