Discover current Kubernetes namespace dinamically
This change removes the hardcoded `toolhive-system` namespace from
configuration and adds a new option `WihCurrentNamespace()` that adds
the namespace written in
`/var/run/secrets/kubernetes.io/serviceaccount/namespace` to the list
of namespaces watched, which currently amounts to just that one
because we do not yet support multi-namespace setups.

Note that this is not only a code quality improvement since it allows
running the Registry Server in namespaces other than `toolhive-system`
while still maintaining its auto-discovery capability.

Fixes #289